+++ controllers/controller.js
@@ -1,7 +1,30 @@
-const queries = require("../models/queries");
+const Book = require("../models/models.js");
 
 exports.createBook = (req, res) => {
-
+// Validate request
+    if (!req.body) {
+        res.status(400).send({
+            message: "Content can not be empty!"
+        });
+    }
+    // Create a Book
+    const book = new Book({
+        isbn: req.body.isbn,
+        title: req.body.title,
+        author: req.body.author,
+        overview: req.body.overview,
+        picture: req.body.picture || false,
+        read_count: req.body.read_count,
+    });
+    // Save Book in the database
+    Book.createBook(book, (err, data) => {
+        if (err)
+            res.status(500).send({
+                message:
+                    err.message || "Some error occurred while creating the book."
+            });
+        else res.send(data);
+    });
 }
 
 exports.findAllBooks = (req, res) => {
